Fixed building of both .jar files in pom.xml . Fixed packename to include 'sc'.
