--- Java-Core/.idea/artifacts/Manager.xml
@@ -1,46 +1,0 @@
-<component name="ArtifactManager">
-  <artifact name="Manager">
-    <output-path>$PROJECT_DIR$/out/artifacts/Manager</output-path>
-    <root id="root">
-      <element id="archive" name="Manager.jar">
-        <element id="directory" name="META-INF">
-          <element id="file-copy" path="$PROJECT_DIR$/src/main/resources/manager-jar/META-INF/MANIFEST.MF" />
-        </element>
-        <element id="module-output" name="SmartRacecar" />
-      </element>
-      <element id="library" level="project" name="Maven: org.w3c:dom:2.3.0-jaxb-1.0.6" />
-      <element id="library" level="project" name="Maven: org.eclipse.paho:org.eclipse.paho.client.mqttv3:1.0.2" />
-      <element id="library" level="project" name="Maven: org.glassfish.jersey.core:jersey-client:2.23.2" />
-      <element id="library" level="project" name="Maven: org.apache.tomcat.embed:tomcat-embed-el:8.5.4" />
-      <element id="library" level="project" name="Maven: org.glassfish.hk2.external:javax.inject:2.5.0-b05" />
-      <element id="library" level="project" name="Maven: org.glassfish.hk2.external:aopalliance-repackaged:2.5.0-b05" />
-      <element id="library" level="project" name="Maven: org.glassfish.jersey.bundles.repackaged:jersey-guava:2.23.2" />
-      <element id="library" level="project" name="Maven: org.javassist:javassist:3.20.0-GA" />
-      <element id="library" level="project" name="Maven: org.apache.tomcat:tomcat-jasper-el:8.5.4" />
-      <element id="library" level="project" name="Maven: org.apache.tomcat.embed:tomcat-embed-core:8.5.4" />
-      <element id="library" level="project" name="Maven: javax.ws.rs:javax.ws.rs-api:2.0.1" />
-      <element id="library" level="project" name="Maven: org.apache.tomcat.embed:tomcat-embed-jasper:8.5.4" />
-      <element id="library" level="project" name="Maven: com.google.code.gson:gson:2.8.0" />
-      <element id="library" level="project" name="Maven: org.glassfish.hk2:hk2-utils:2.5.0-b05" />
-      <element id="library" level="project" name="Maven: org.apache.tomcat:tomcat-servlet-api:8.5.4" />
-      <element id="library" level="project" name="Maven: org.apache.tomcat:tomcat-el-api:8.5.4" />
-      <element id="library" level="project" name="Maven: org.apache.tomcat:tomcat-jsp-api:8.5.4" />
-      <element id="library" level="project" name="Maven: org.glassfish.hk2:hk2-locator:2.5.0-b05" />
-      <element id="library" level="project" name="Maven: org.apache.tomcat:tomcat-jasper:8.5.4" />
-      <element id="library" level="project" name="Maven: javax.validation:validation-api:1.1.0.Final" />
-      <element id="library" level="project" name="Maven: org.eclipse.jdt.core.compiler:ecj:4.5.1" />
-      <element id="library" level="project" name="Maven: org.glassfish.hk2:osgi-resource-locator:1.0.1" />
-      <element id="library" level="project" name="Maven: org.apache.tomcat:tomcat-util:8.5.4" />
-      <element id="library" level="project" name="Maven: org.apache.tomcat:tomcat-api:8.5.4" />
-      <element id="library" level="project" name="Maven: org.glassfish.jersey.media:jersey-media-jaxb:2.23.2" />
-      <element id="library" level="project" name="Maven: org.glassfish.jersey.core:jersey-common:2.23.2" />
-      <element id="library" level="project" name="Maven: org.glassfish.hk2:hk2-api:2.5.0-b05" />
-      <element id="library" level="project" name="Maven: org.apache.tomcat:tomcat-juli:8.5.4" />
-      <element id="library" level="project" name="Maven: org.apache.tomcat:tomcat-util-scan:8.5.4" />
-      <element id="library" level="project" name="Maven: org.glassfish.jersey.containers:jersey-container-servlet-core:2.23.2" />
-      <element id="library" level="project" name="Maven: org.glassfish.jersey.containers:jersey-container-servlet:2.23.2" />
-      <element id="library" level="project" name="Maven: org.glassfish.jersey.core:jersey-server:2.23.2" />
-      <element id="library" level="project" name="Maven: javax.annotation:javax.annotation-api:1.2" />
-    </root>
-  </artifact>
-</component>
