+++ Java-Core/release/Core/maps/V314.yaml
@@ -1,0 +1,7 @@
+image: V314.pgm
+resolution: 0.050000
+origin: [-12.824999, -12.824999, 0.000000]
+negate: 0
+occupied_thresh: 0.65
+free_thresh: 0.196
+
