+++ Java-Core/release/Core/maps/maps.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<maps>
+	<map>
+		<name>V314</name>
+	</map>
+	<map>
+		<name>z164</name>
+	</map>
+<map>
+    <name>zbuilding</name>
+  </map>
+</maps>
