+++ Java-Core/src/main/java/be/uantwerpen/fti/ds/sc/smartracecar/common/MQTTListener.java
@@ -1,0 +1,6 @@
+package be.uantwerpen.fti.ds.sc.smartracecar.common;
+
+//All MQTT functionality
+public interface MQTTListener {
+    void parseMQTT(String topic, String message);
+}
