+++ Java-Core/src/main/java/be/uantwerpen/fti/ds/sc/smartracecar/manager/JerseyConfiguration.java
@@ -1,0 +1,10 @@
+package be.uantwerpen.fti.ds.sc.smartracecar.manager;
+
+import org.glassfish.jersey.server.ResourceConfig;
+
+public class JerseyConfiguration extends ResourceConfig {
+
+    public JerseyConfiguration() {
+        packages("be.uantwerpen.fti.ds.sc.smartracecar.manager");
+    }
+}
