--- Java-Core/src/main/java/be/uantwerpen/fti/ds/smartracecar/common/MQTTListener.java
@@ -1,6 +1,0 @@
-package be.uantwerpen.fti.ds.smartracecar.common;
-
-//All MQTT functionality
-public interface MQTTListener {
-    void parseMQTT(String topic, String message);
-}
