--- Java-Core/src/main/java/be/uantwerpen/fti/ds/smartracecar/common/Map.java
@@ -1,11 +1,0 @@
-package be.uantwerpen.fti.ds.smartracecar.common;
-
-//common of a offline kept map.
-public class Map {
-
-    private String name = ""; //Name of the mapfile.
-
-    public Map(String name) {
-        this.name = name;
-    }
-}
