--- Java-Core/src/main/java/be/uantwerpen/fti/ds/smartracecar/common/WayPoint.java
@@ -1,18 +1,0 @@
-package be.uantwerpen.fti.ds.smartracecar.common;
-
-//Extended common of a Point which is used for waypoints. Supers the coordinates from the Point class.
-public class WayPoint extends Point {
-
-    private int ID = 0; // waypoint ID given by RaceCarManager.
-    private int weight = 0; //weight used for pathfinding algorthims.
-
-    public WayPoint(int ID,float x, float y,float z, float w, int weight) {
-        super(x, y, z, w);
-        this.ID = ID;
-        this.weight = weight;
-    }
-
-    public int getID() {
-        return ID;
-    }
-}
