--- Java-Core/src/main/java/be/uantwerpen/fti/ds/smartracecar/manager/JerseyConfiguration.java
@@ -1,10 +1,0 @@
-package be.uantwerpen.fti.ds.smartracecar.manager;
-
-import org.glassfish.jersey.server.ResourceConfig;
-
-public class JerseyConfiguration extends ResourceConfig {
-
-    public JerseyConfiguration() {
-        packages("be.uantwerpen.fti.ds.smartracecar.manager");
-    }
-}
