--- Java-Core/src/main/resources/manager-jar/META-INF/MANIFEST.MF
@@ -1,16 +1,0 @@
-Manifest-Version: 1.0
-Class-Path: dom-2.3.0-jaxb-1.0.6.jar org.eclipse.paho.client.mqttv3-1.
- 0.2.jar jersey-client-2.23.2.jar tomcat-embed-el-8.5.4.jar javax.inje
- ct-2.5.0-b05.jar aopalliance-repackaged-2.5.0-b05.jar jersey-guava-2.
- 23.2.jar javassist-3.20.0-GA.jar tomcat-jasper-el-8.5.4.jar tomcat-em
- bed-core-8.5.4.jar javax.ws.rs-api-2.0.1.jar tomcat-embed-jasper-8.5.
- 4.jar gson-2.8.0.jar hk2-utils-2.5.0-b05.jar tomcat-servlet-api-8.5.4
- .jar tomcat-el-api-8.5.4.jar tomcat-jsp-api-8.5.4.jar hk2-locator-2.5
- .0-b05.jar tomcat-jasper-8.5.4.jar validation-api-1.1.0.Final.jar ecj
- -4.5.1.jar osgi-resource-locator-1.0.1.jar tomcat-util-8.5.4.jar tomc
- at-api-8.5.4.jar jersey-media-jaxb-2.23.2.jar jersey-common-2.23.2.ja
- r hk2-api-2.5.0-b05.jar tomcat-juli-8.5.4.jar tomcat-util-scan-8.5.4.
- jar jersey-container-servlet-core-2.23.2.jar jersey-container-servlet
- -2.23.2.jar jersey-server-2.23.2.jar javax.annotation-api-1.2.jar
-Main-Class: be.uantwerpen.fti.ds.smartracecar.manager.Manager
-
